Dobra działa.

+++ src/visualisations/Transporters/VisualisationTransportersModel.java
@@ -1,6 +1,7 @@
 package visualisations.Transporters;
 
 import java.awt.Color;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Random;
@@ -28,6 +29,8 @@ import dataModels.Transporter;
 import dataModels.User;
 import database.DataAccessObjectFactory;
 import database.DataAccessObjectTransportersVisualisation;
+
+
 
 /**
  * Model wizualizacji przewoźników
@@ -146,17 +149,17 @@ public class VisualisationTransportersModel
 
 		
 		for(ArrayList<Object> row : result_small){
-			transporters.add(new Transporter((int)row.get(0), a, (int)row.get(1), (double)row.get(2), (int)row.get(3), (int)row.get(4), (double)row.get(8), (double)row.get(7), (String)row.get(5), (int)row.get(6)));
+			transporters.add(new Transporter((int)row.get(0), a, ((Long)row.get(1)).intValue(), (double)row.get(2), ((BigDecimal)row.get(3)).intValue(), ((BigDecimal)row.get(4)).intValue(), ((BigDecimal)row.get(8)).intValue(), ((BigDecimal)row.get(7)).intValue(), (String)row.get(5), (int)row.get(6)));
 		}
+			
 		
 		for(ArrayList<Object> row : result_medium){
-			transporters.add(new Transporter((int)row.get(0), b, (int)row.get(1), (double)row.get(2), (int)row.get(3), (int)row.get(4), (double)row.get(8), (double)row.get(7), (String)row.get(5), (int)row.get(6)));
-		}
+			transporters.add(new Transporter((int)row.get(0), b, ((Long)row.get(1)).intValue(), (double)row.get(2), ((BigDecimal)row.get(3)).intValue(), ((BigDecimal)row.get(4)).intValue(), ((BigDecimal)row.get(8)).intValue(), ((BigDecimal)row.get(7)).intValue(), (String)row.get(5), (int)row.get(6)));
+			}
 		
 		for(ArrayList<Object> row : result_big){
-			transporters.add(new Transporter((int)row.get(0), c, (int)row.get(1), (double)row.get(2), (int)row.get(3), (int)row.get(4), (double)row.get(8), (double)row.get(7), (String)row.get(5), (int)row.get(6)));
-		}
-		
+			transporters.add(new Transporter((int)row.get(0), c, ((Long)row.get(1)).intValue(), (double)row.get(2), ((BigDecimal)row.get(3)).intValue(), ((BigDecimal)row.get(4)).intValue(), ((BigDecimal)row.get(8)).intValue(), ((BigDecimal)row.get(7)).intValue(), (String)row.get(5), (int)row.get(6)));
+			}
 		
 		
 		/*
